Specify sane defaults for demo certificates
Previously, certificate generation expected access to S3.

This shouldn't be required when generating a demo certificate, so we
update the README's sample script to disable S3 uploads.

The sample script now specifies the default demo course_id, instead of a
non-existent course_id.

All command line flags are specified with `--long-options` to make the
options explicit to users, as this is their first exposure to the code.