Remove over-cautious input validation
This already has a sane default (`INFO`), so we don't gain anything by
silenting overwriting an invalid configuration. Instead, we should pass
along any developer-provided log level. If it causes an error with the
logging framework, it's advantageous to have it break immediately upon
an invalid configuration. This forces the developer to address the issue
immediately.