Refactor CME rendering
* DO NOT BLINDLY ADOPT THIS PATTERN IN FUTURE USE! *

This is _not_ meant to be a long-term re-architecting; it's more of a
stop-gap code clean up. This repo is littered with cargo cult and
technical debt which has amassed over the years. The `gen_cert.py` is
too large and contains too much code.

In the future, I'd like to see _all_ rendering functionality moved into
external, PIP-installable package repositories. A developer should be
able to `pip install` a renderer, having it become automatically
available to the underlying certificate generation framework.

In the meantime, however, we can at least begin moving shared and
instance-specific out into separate files/directories.