Cast fontlist `generator` back to a `list`
This _should_ be a generator, but we overlooked that one line of code
attempts to index directly into the generator, as if it were a `dict`.

Ideally, we should rework the logic to support the generator, but
frankly, we needed a quick fix and this is the least intensive path
forward.

TODO: Fix this.